benchtests: Avoid truncation in random memcpy/memset benchmarks
Use uint16_t rather than uint8_t for the size arrays.

+++ benchtests/bench-memset-random.c
@@ -39,7 +39,7 @@ typedef struct { uint8_t align; uint16_t freq; } align_data_t;
 
 #define SIZE_NUM 65536
 #define SIZE_MASK (SIZE_NUM-1)
-static uint8_t len_arr[SIZE_NUM];
+static uint16_t len_arr[SIZE_NUM];
 
 /* Frequency data for memset sizes up to 4096 bytes based on SPEC2017.  */
 static freq_data_t memset_len_freq[] =
